Move all entity simulation logic to server classes
